+++ raven_front/lib/pages/manage/reissue/restricted.dart
@@ -29,7 +29,7 @@ class _ReissueRestrictedAssetState extends State<ReissueRestrictedAsset> {
       streams.reissue.form.add(GenericReissueForm(
         parent: asset.parent?.symbol,
         name: asset.symbol,
-        minQuantity: asset.amount.toInt(),
+        minQuantity: asset.amount,
         minDecimal: asset.divisibility,
         decimal: asset.divisibility,
         minIpfs: asset.ipfs,
